+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## v2.3.10 (1.21)
+### Additions
+- added compat for LambDynamicLights
 ### Changes
 ### Fixes